requirements.txt
azure-functions==1.0.7
jira==2.0.0
jira-python==0.2.2
